pyproject.toml
# LLM-Enhanced MARL-Based IDS for IoT
# Project Configuration

[project]
name = "llm-marl-ids"
version = "1.0.0"
description = "LLM-Enhanced Multi-Agent Reinforcement Learning Intrusion Detection System for IoT"
authors = ["Your Name <your.email@institution.edu>"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"

[project.urls]
Homepage = "https://github.com/yourusername/llm-marl-ids"
Documentation = "https://llm-marl-ids.readthedocs.io"
Repository = "https://github.com/yourusername/llm-marl-ids"

[tool.poetry]
name = "llm-marl-ids"
version = "1.0.0"
description = "LLM-Enhanced MARL IDS for IoT"
authors = ["Your Name <your.email@institution.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
transformers = "^4.35.0"
sentence-transformers = "^2.2.2"
ray = {extras = ["rllib"], version = "^2.8.0"}
scapy = "^2.5.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
wandb = "^0.16.0"
tensorboard = "^2.15.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
openai = "^1.3.0"
grpcio = "^1.59.0"
paho-mqtt = "^1.6.1"
h5py = "^3.10.0"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.10.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"
ipython = "^8.16.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
